requirements.txt
flask-SQLALCHEMY==2.2
marshmallow==2.13.1
requests
discogs_client
beautifulsoup4